+++ src/main/java/top/warmj/pojo/Result.java
@@ -3,7 +3,7 @@ package top.warmj.pojo;
 import java.io.Serializable;
 
 public class Result<T> implements Serializable {
-    private static final long serialVersionUID = 1L;
+    private static final int serialVersionUID = 1;
 
     public static final int SUCCESS = 0;
